商品导入-新增和sku、规格组、规格键值校验完成

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/goods/goodsimport/vpu/GoodsVpuExcelImportBo.java
@@ -25,7 +25,7 @@ public class GoodsVpuExcelImportBo extends GoodsExcelImportBase {
     private BigDecimal shopPrice;
     private Byte isOnSale;
     private Integer limitBuyNum;
-    private Integer goodsWeight;
+    private BigDecimal goodsWeight;
     private String unit;
     /**商品图片使用;分隔*/
     private String goodsImgsStr;
